#2 restructured classes
